Extract GOVUK support into razor class library NIHR.GovUk.AspNetCore.Mvc
