+++ BPOR/BPOR.Rms/Controllers/ResearcherController.cs
@@ -5,6 +5,7 @@ using BPOR.Rms.Models.Study;
 using BPOR.Rms.Startup;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using NIHR.GovUk.AspNetCore.Mvc;
 
 namespace BPOR.Rms.Controllers;
