+++ BPOR/BPOR.Rms/Controllers/StudyController.cs
@@ -4,6 +4,7 @@ using BPOR.Rms.Models.Study;
 using BPOR.Rms.Startup;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using NIHR.GovUk.AspNetCore.Mvc;
 using NIHR.Infrastructure.Paging;
 
 namespace BPOR.Rms.Controllers;
