+++ BPOR/BPOR.Rms/Views/Email/SetupCampaign.cshtml
@@ -1,5 +1,5 @@
 @using BPOR.Rms.Models.Email
-@model BPOR.Rms.Models.Email.SetupCampaignViewModel
+@model SetupCampaignViewModel
 
 @{
     ViewData["Title"] = "Set up an email campaign";
