+++ BPOR/BPOR.Rms/Views/Shared/_Layout.cshtml
@@ -35,8 +35,8 @@
         <back-link />
         <breadcrumbs />
         <main class="govuk-main-wrapper" id="main-content" role="main">
-            <partial name="Partials/_NotificationBanner" />
-            <partial name="Partials/_ErrorSummary" />
+            <partial name="_NotificationBanner" />
+            <partial name="_ErrorSummary" />
             @RenderBody()
         </main>
     </div>
