+++ BPOR/BPOR.Rms/Views/_ViewImports.cshtml
@@ -1,5 +1,7 @@
 ﻿@using BPOR.Rms
 @using BPOR.Rms.Models
+@using NIHR.GovUk.AspNetCore.Mvc
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, BPOR.Rms
 @addTagHelper *, BPOR.Rms.TagHelpers
+@addTagHelper *, NIHR.GovUk.AspNetCore.Mvc
