+++ BPOR/NIHR.GovUk.AspNetCore.Mvc/TagHelpers/FormGroupTagHelper.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 using System.Text.Encodings.Web;
 
-namespace BPOR.Rms.TagHelpers
+namespace NIHR.GovUk.AspNetCore.Mvc.TagHelpers
 {
     public class FormGroupTagHelper : TagHelper
     {
