+++ BPOR/NIHR.GovUk.AspNetCore.Mvc/Views/Shared/_Breadcrumbs.cshtml
@@ -1,3 +1,4 @@
+@using NIHR.GovUk.AspNetCore.Mvc
 @{
     var breadcrumbs = ViewData["Breadcrumbs"] as IList<BreadcrumbItem>;
 }
