+++ BPOR/NIHR.GovUk.AspNetCore.Mvc/Views/Shared/_NotificationBanner.cshtml
@@ -1,4 +1,5 @@
-@using NIHR.Infrastructure.AspNetCore;
+@using NIHR.Infrastructure.AspNetCore
+@using NIHR.GovUk.AspNetCore.Mvc
 <style>
     .govuk-notification-banner:focus {
         outline: 3px solid #FED47A
